@@ -5,8 +5,6 @@ namespace Database\Factories;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
-use App\Models\User;
-use Faker\Factory as FakerFactory;
 
 /**
  * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\User>
@@ -17,7 +15,6 @@ class UserFactory extends Factory
      * The current password being used by the factory.
      */
     protected static ?string $password;
-    protected $model = User::class;
 
     /**
      * Define the model's default state.
@@ -26,16 +23,12 @@ class UserFactory extends Factory
      */
     public function definition(): array
     {
-        $faker = FakerFactory::create('en_US');
         return [
-            'user_role' => $faker->randomElement(['admin', 'user', 'doctor']),
-            'username' => $faker->userName,
-            'password' => Hash::make('password'),
-            'nama' => $faker->name,
-            'email' => $faker->unique()->safeEmail,
-            'telepon' => $faker->phoneNumber,
-            'alamat' => $faker->address,
-            'gender' => $faker->randomElement(['male', 'female']),
+            'name' => fake()->name(),
+            'email' => fake()->unique()->safeEmail(),
+            'email_verified_at' => now(),
+            'password' => static::$password ??= Hash::make('password'),
+            'remember_token' => Str::random(10),
         ];
     }
 
