@@ -1,40 +1,22 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\MqttController;
-use App\Http\Controllers\AuthController;
+use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\AdminController;
-use App\Http\Controllers\HomeController;
-use App\Http\Controllers\CalculationController;
+use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "web" middleware group. Make something great!
-|
-*/
+Route::get('/', function () {
+    return view('welcome');
+});
+
 
 Route::controller(AdminController::class, '')->group(function(){
-    Route::get('/admin-dashboard', 'index')->name('admin-dashboard');
+    Route::get('/dashboard', 'index')->middleware(['auth', 'verified'])->name('dashboard');
 });
 
-Route::get('/home', [HomeController::class, 'index'])->name('home');
-
-Route::controller(AuthController::class, '')->group(function(){
-    Route::post('/login', 'login')->name('login');
-    Route::get('/login', 'loginView')->name('loginView');
-
-    Route::post('/register', 'register')->name('register');
-    Route::get('/register', 'registerView')->name('registerView');
+Route::middleware('auth')->group(function () {
+    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
+    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::controller(MqttController::class, '')->group(function(){
-    Route::post('/publish', 'publishMessage')->name('publish');
-    Route::get('/subscribe', 'subscribe')->name('subscribe');
-});
-
-Route::get('/calculate', [CalculationController::class, 'calculate'])->name('calculate');
+require __DIR__.'/auth.php';
